Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/aborted-parser.window-expected.txt


PASS document.open() after parser is aborted
PASS async document.open() after parser is aborted
